requirements.txt
# --- Base ---
Django==6.0.2
gunicorn==25.1.0
psycopg2-binary==2.9.11
dj-database-url==3.1.2
whitenoise==6.12.0
pillow==12.0.0

# --- Interface & Admin ---
django-jazzmin==3.0.4
django-ckeditor-5==0.2.20

# --- Formulaires (Ce qui manque ici) ---
django-crispy-forms==2.3
crispy-bootstrap4==2024.10

# --- Email ---
django-anymail==12.0

# --- Traduction & Apps ---
django-modeltranslation==0.19.11
django_moncash==2.6.0
moncash==1.0.3
stripe==14.4.1

# --- API REST ---
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
django-filter==25.2
PyJWT==2.12.1

# --- Utilitaires ---
python-slugify==8.0.4
requests==2.32.5
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3